Cleanup

+++ app/src/main/java/com/example/cryptomanager_v2/data/model/ExchangeRates/ExchangeRatesOld.kt
@@ -30,7 +30,7 @@ class ExchangeRatesOld {
                     fiatsArray.add(
                         DBFiat(
                             name = key,
-                            rate = value.toDouble()
+                            rate = value
                         )
                     )
                 }
